change location of func
